refactor(index screen): render overview components conditionally
- duplicated index screen to two components, one for event records
  and one for the rest
- updated `IndexScreen` to render the right component conditionally

SVA-1114
