+++ src/components/screens/index.js
@@ -3,10 +3,12 @@ export * from './consul';
 export * from './About';
 export * from './AvailableVehicles';
 export * from './EventRecord';
+export * from './EventRecords';
 export * from './LunchSection';
 export * from './NewsItem';
 export * from './OpeningTimesCard';
 export * from './OperatingCompany';
+export * from './Overviews';
 export * from './PointOfInterest';
 export * from './PriceCard';
 export * from './Service';
